pyproject.toml
[tool.poetry]
name = "athena"
version = "0.1.0"
description = "Backend part of athena system"
authors = ["RTU MIREA IPPO Laboratory"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
django-dotenv = "^1.4"
django-cors-headers = "^2.4"
djangorestframework = "^3.9"
djangorestframework_simplejwt = "^3.3"
six = "^1.12"
djangorestframework-camel-case = "^1.0"
gunicorn = "^19.9"
psycopg2-binary = "^2.7"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0",allows-prereleases = true}
isort = "^4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
